Extract Google LLM base entity class
